[WIP] refactor existing code into preload and clms classes
